requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
fa2==0.3.5
Flask==2.2.2
fonttools==4.38.0
gensim==4.2.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonlines==3.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
networkx==2.8.8
numpy==1.23.5
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-louvain==0.16
pytz==2022.6
requests==2.28.1
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
smart-open==6.2.0
tqdm==4.64.1
urllib3==1.26.12
Werkzeug==2.2.2
wordcloud==1.8.2.2
zipp==3.10.0
